Creating the functionalities to register a note and display registered notes
Creating the functionality to register notes, by creating a new controller, the 'NotesController', with the 'create' and 'show' methods, in order to save a note in the database and display note information, respectively. Also creating the 'notes.routes.js' file to manage routes related to modifying and registering notes. Lastly, addressing some database errors and fixing file names