Creating the functionalities to delete a note and obtain a list of notes
Adding the functionality to delete a recorded note through the 'delete' method of the 'NotesController' class, identifying which note will be deleted by the provided note id in the request as a parameter. Finally, implementing the feature to display a list of notes with various filtering options for the list, either by note tag, note title, or the user id who created the notes, through the 'index' method.