Starting the user authentication logic, through the SessionsController and the /sessions route
Starting the construction of the code responsible for user authentication, creating the 'SessionsController' controller responsible for user validation and token authentication. Also, creating the 'sessions.routes.js' file to manage authentication routes. Establishing the '/sessions' route in the 'index.jsx' file within the 'routes' folder, and creating the 'configs' folder along with the 'auth.js' file to store the necessary configurations for generating the authentication token. Finally, installing the 'jsonwebtoken' library to generate the token used in the application.