Completing the authentication logic by creating the 'ensureAuthenticated' middleware
Concluding the development of the code responsible for user authentication, creating the 'middlewares' folder and the 'ensureAuthenticated.js' middleware that will intercept requests and verify if the authentication token has been provided, throwing an error if authentication is not satisfied. If authentication is successful and there is a token present, the controller responsible for the request will be called, receiving the user's ID in the request body. Modifying the routes within the 'routes' folder to implement the 'ensureAuthenticated' middleware. Finally, updating the application's controllers to receive the user's ID in the request body instead of as a parameter