Developing the code responsible for uploading the user's avatar image
Creating the logic and necessary code to perform the upload of the user's avatar image and save the image information along with other user details. Creating the 'upload.js' file within the 'config' folder, which stores the necessary settings for image uploads. Creating the 'tmp' folder at the root of the project to temporarily store the image uploaded by the user and creating the 'uploads' folder within the 'tmp' folder where the image will be permanently stored. Installing the 'multer' library to handle the images that will be sent by the user via HTTP requests. Creating the 'providers' folder and the 'DiskStorage' class containing the 'saveFile' and 'deleteFile' methods responsible for saving the image in the correct folder and deleting the image, respectively. Creating the '/avatar' route within the 'users.routes.js' file, responsible for managing user routes, where the image will be sent to update the user's profile picture. Using the 'multer' library, specifying the path where the user's avatar photo will be stored, to intercept the request, change the file name, and save it in the 'tmp' folder. Finally, creating the 'UserAvatarController', responsible for receiving the request, saving the image in the 'uploads' folder using the methods of the 'DiskStorage' class, and saving the image name in the database using the 'avatar' field