Preparing the project for deployment by installing the 'dotenv' and 'pm2' dependencies.
Preparing the application for deployment. Installing the 'dotenv' library to handle sensitive data in the application by storing these data in environment variables in the production environment. Creating the '.env.example' file, containing the structure of the variables being used but without the values. Finally, installing the 'pm2' library and generating the 'ecosystem.config.cjs' file to ensure that in the production environment, the application continues to run intermittently.