Preparing the application for the implementation of automated tests using the 'jest' framework
Updating the application by installing the 'jest' framework for the creation and implementation of automated tests. Creating the 'jest.config.js' file containing the necessary configurations for the 'jest' to function and specifying which files will be the test files. Generating the 'mytest.spec.js' file with a sample test to understand how the creation and execution of tests work. Lastly, further dividing the application's responsibilities by creating the 'repositories' folder and the 'UserRepository.js' file to encapsulate the logic of communication with the user table in the database through the 'findByEmail' and 'create' methods