Implementing 'UserRepository' in 'UsersController' and applying the principle of dependency inversion in the application
Using the 'UserRepository' class, which contains the logic for communicating with the database, in the 'UsersController' to increase the division of responsibilities and reduce coupling. Applying the principle of dependency inversion to the application's classes by creating the 'services' folder along with the 'UserCreateService.js' file to encapsulate the logic and responsibility for user creation. Finally, modifying the 'UsersController' to be responsible only for receiving the request and sending the response to the user. The logic for user creation is assigned to the 'UserCreateService' class, and the communication with the database is assigned to the 'UserRepository' class