Creating a test file 'UserCreateService.spec.js' to test the logic of user creation and creating the 'UserRepositoryInMemory' class to simulate the database connection for the test environment.
Implementing automated tests in the application. Removing the example file 'mytest.spec.js' and creating the test file 'UserCreateService.spec.js' within the 'services' folder, with the goal of testing the user registration logic present in the 'UserCreateService' class. Updating the 'UserCreateService' class to return the value returned by the 'create' method of the 'userRepository' and removing unnecessary imports and variables. Creating the 'UserRepositoryInMemory' class to simulate communication with the user table in the database, present in the 'userRepository', since for testing, one should not use or depend on external resources such as a database connection. Modifying the 'jest' script in the 'package.json' file to enable the use of ES6 modules in test files and using the '--watch' flag to continuously run the tests, rerunning the test each time a new change occurs. Finally, creating tests in the 'UserCreateService.spec.js' file to verify if the user creation logic is working correctly and checking if an error is triggered when attempting to register a new user with an existing email